+++ src/Map.h
@@ -35,7 +35,7 @@ public:
      * @param timestamp timestamp to be inserted
      * @return true for success, false for failure
      */
-    bool put(const char *key, size_t keylen, const char *value, size_t valuelen, uint64_t timestamp);
+    bool put(const char *key, uint32_t keylen, const char *value, uint32_t valuelen, uint64_t timestamp);
 
     /**
      * get the value for a specific key
@@ -71,7 +71,7 @@ public:
     /**
      * size of KV when serialized and added to map
      */
-    static size_t kv_size(const char *key, size_t keylen, const char *value, size_t valuelen, uint64_t timestamp);
+    static uint32_t kv_size(const char *key, uint32_t keylen, const char *value, uint32_t valuelen, uint64_t timestamp);
 
     /**
      * clear all elements of map
