+++ src/MemStore.h
@@ -44,7 +44,7 @@ public:
      * @param timestamp timestamp to be inserted
      * @return true for success, false for failure
      */
-    bool put(const char *key, size_t keylen, const char *value, size_t valuelen, uint64_t timestamp);
+    bool put(const char *key, uint32_t keylen, const char *value, uint32_t valuelen, uint64_t timestamp);
 
     /**
      * get the (copy of) value for a specific key.
@@ -75,7 +75,7 @@ public:
      * check if adding <key, value, timestamp> to memstore will reach memstore's
      * size limit
      */
-    bool will_reach_size_limit(const char *key, size_t keylen, const char *value, size_t valuelen, uint64_t timestamp);
+    bool will_reach_size_limit(const char *key, uint32_t keylen, const char *value, uint32_t valuelen, uint64_t timestamp);
 
     /**
      * clear memstore
